Revert "Database update (#28)"
This reverts commit 10fe973f44c8b27bca9cceb915f4147d86c599fc.
