+++ app/build.gradle
@@ -33,8 +33,12 @@ dependencies {
     implementation 'com.google.firebase:firebase-storage:16.0.4'
     //implementation 'com.google.firebase:firebase-crash:16.0.4'
     implementation 'com.google.firebase:firebase-auth:16.0.4'
+
     implementation 'com.firebase:firebase-client-android:2.3.1'
+
     implementation 'com.squareup.picasso:picasso:2.5.2'
+
+
     implementation fileTree(include: ['*.jar'], dir: 'libs')
     implementation 'com.android.support:appcompat-v7:28.0.0'
     implementation 'com.android.support:design:28.0.0'
@@ -48,6 +52,7 @@ dependencies {
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
     //recyclerView
     implementation 'com.android.support:recyclerview-v7:28.0.0'
+
     //https://github.com/bumptech/glide
     implementation 'com.github.bumptech.glide:glide:4.8.0'
     annotationProcessor 'com.github.bumptech.glide:compiler:4.8.0'
