+++ app/src/main/java/gcsenxmk/q/FirebaseLoginActivity.java
@@ -11,16 +11,10 @@ import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
 
-import com.firebase.client.DataSnapshot;
-import com.firebase.client.Firebase;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.database.DatabaseError;
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.database.FirebaseDatabase;
-import com.google.firebase.database.ValueEventListener;
 
 public class FirebaseLoginActivity extends AppCompatActivity {
 
@@ -31,7 +25,12 @@ public class FirebaseLoginActivity extends AppCompatActivity {
 
     private ProgressDialog progressDialog;
     private FirebaseAuth firebaseAuth;
-  
+
+
+
+
+
+
     @Override
     protected void onCreate(Bundle savedInstanceState){
 
@@ -75,6 +74,9 @@ public class FirebaseLoginActivity extends AppCompatActivity {
             }
         });
 
+
+
+
 }
 
     private  void userLogin(){
@@ -110,6 +112,12 @@ public class FirebaseLoginActivity extends AppCompatActivity {
         });
 
     }}
+
+
+
+
+
+
 
 
 //    private void signIn(final String username, final String password) {
