+++ app/src/main/java/gcsenxmk/q/MerchantViewQueuesActivityAfterLogin.java
@@ -1,16 +1,13 @@
 package gcsenxmk.q;
 
-import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
-import android.widget.EditText;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.firebase.client.collection.LLRBEmptyNode;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.database.DataSnapshot;
@@ -29,7 +26,7 @@ public class MerchantViewQueuesActivityAfterLogin extends AppCompatActivity{
     public int len;
     QueueInformation queueInformation;
 
-    private DatabaseReference queueDatabaseRef,customerDatabaseReference, merchantDatabaseReference;
+    private DatabaseReference queueDatabaseRef;
 
     @Override
 
@@ -46,8 +43,7 @@ public class MerchantViewQueuesActivityAfterLogin extends AppCompatActivity{
         btnCreateQueue= (Button) findViewById(R.id.btnCreateNewQueue);
 
         queueDatabaseRef = FirebaseDatabase.getInstance().getReference("Queue");
-        customerDatabaseReference=FirebaseDatabase.getInstance().getReference("Users");
-        merchantDatabaseReference=FirebaseDatabase.getInstance().getReference("Merchants");
+
 
         queueDatabaseRef.child(user.getUid()).addListenerForSingleValueEvent(new ValueEventListener() {
             @Override
@@ -55,7 +51,6 @@ public class MerchantViewQueuesActivityAfterLogin extends AppCompatActivity{
                 for (DataSnapshot childSnapshot : dataSnapshot.getChildren()) {
                     queueInformation = dataSnapshot.getValue(QueueInformation.class);
                     len = queueInformation.queue.size();
-                    //System.out.println(len);
                     queue_length.setText(Integer.toString(len));
                 }
             }
@@ -71,36 +66,8 @@ public class MerchantViewQueuesActivityAfterLogin extends AppCompatActivity{
             @Override
             public void onClick(View v) {
                 if(len>=1) {
-                    System.out.println(queueInformation.queue.size());
-                    String current_user_in_queue= queueInformation.queue.get(0);
                     queueInformation.queue.remove(0);
-                    System.out.println(current_user_in_queue);
-                    customerDatabaseReference.addListenerForSingleValueEvent(new ValueEventListener() {
-                        @Override
-                        public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
-                            for(DataSnapshot childSnapshot:dataSnapshot.getChildren()){
-                                //System.out.println("yo"+childSnapshot);
-                                if(childSnapshot.getKey().equals(current_user_in_queue)) {
-                                   // System.out.println(childSnapshot.child(current_user_in_queue));
-                                    System.out.println(customerDatabaseReference.child(childSnapshot.getKey().toString()).child("merchantID").toString());
-                                    customerDatabaseReference.child(childSnapshot.getKey().toString()).child(("merchantID").toString()).removeValue();
-
-                                    //customerDatabaseReference.child(childSnapshot.child(current_user_in_queue).toString()).child(("merchantID").toString()).removeValue();
-                                    //queueDatabaseRef.child(user.getUid()).removeValue();
-                                }else{
-                                    System.out.println("Error");
-                                }
-
-                            }
-                        }
-
-                        @Override
-                        public void onCancelled(@NonNull DatabaseError databaseError) {
-
-                        }
-                    });
-                    Toast.makeText(MerchantViewQueuesActivityAfterLogin.this, "finish the service for onen customer", Toast.LENGTH_SHORT).show();
-
+                    Toast.makeText(MerchantViewQueuesActivityAfterLogin.this, "fiish the service for onen customer", Toast.LENGTH_SHORT).show();
                     queueDatabaseRef.child(user.getUid()).child("queue").setValue(queueInformation.queue);
                     len--;
                     queue_length.setText(Integer.toString(len));
@@ -132,9 +99,7 @@ public class MerchantViewQueuesActivityAfterLogin extends AppCompatActivity{
                             Toast.makeText(MerchantViewQueuesActivityAfterLogin.this,
                                     "You already have a queue", Toast.LENGTH_LONG).show();
                         }else{
-                            Intent intent= new Intent(MerchantViewQueuesActivityAfterLogin.this, CreateQueueUponLoginActivity.class);
-                            startActivity(intent);
-                            //Toast.makeText(MerchantViewQueuesActivityAfterLogin.this,"Add queue now", Toast.LENGTH_SHORT).show();
+                            Toast.makeText(MerchantViewQueuesActivityAfterLogin.this,"Add queue now", Toast.LENGTH_SHORT).show();
                         }
 
                     }
