+++ app/src/main/java/gcsenxmk/q/ProfileActivity.java
@@ -13,7 +13,6 @@ import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
-import com.google.firebase.iid.FirebaseInstanceIdService;
 
 public class ProfileActivity extends AppCompatActivity {
     private FirebaseAuth firebaseAuth;
@@ -23,6 +22,11 @@ public class ProfileActivity extends AppCompatActivity {
 
     private EditText name,area;
     private Button save;
+
+
+
+
+
 
     @Override
