+++ app/src/main/java/gcsenxmk/q/QueueInformation.java
@@ -5,26 +5,23 @@ import android.graphics.Bitmap;
 import java.util.ArrayList;
 
 public class QueueInformation {
-    public String queuename;
-    public String desc;
-    public int average_waiting_time;
 
-    public static ArrayList<String> queue;
+    public static int creation = 0;
+    public String queuename;
+
+    public ArrayList<String> queue;
 
     public QueueInformation(){
         this.queuename = "";
-        this.desc="";
-        this.average_waiting_time=1;
         queue = new ArrayList<String>();
-
+       // this.queue.add("FirstUser");
 
     }
 
-    public QueueInformation(String name, String desc, int average_waiting_time) {
+    public QueueInformation(String name) {
         this.queuename = name;
-        this.desc=desc;
-        this.average_waiting_time=average_waiting_time;
         queue = new ArrayList<String>();
+        //this.queue.add("initial");
 
     }
 
@@ -40,22 +37,6 @@ public class QueueInformation {
 
     public ArrayList<String> getQueue() {
         return queue;
-    }
-
-    public String getDesc() {
-        return desc;
-    }
-
-    public void setDesc(String desc) {
-        this.desc = desc;
-    }
-
-    public int getWaiting_time_customer() {
-        return average_waiting_time;
-    }
-
-    public void setWaiting_time_customer(int waiting_time_customer) {
-        this.average_waiting_time = waiting_time_customer;
     }
 
     public void setQueue(ArrayList<String> queue) {
