+++ app/src/main/java/gcsenxmk/q/RecyclerActivity.java
@@ -3,16 +3,12 @@ package gcsenxmk.q;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
+
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
-import android.view.View;
-import android.widget.ImageButton;
 import android.widget.ProgressBar;
 import android.widget.Toast;
 
-import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
@@ -42,7 +38,7 @@ public class RecyclerActivity extends AppCompatActivity {
         setContentView(R.layout.activity_recycler);
 
         mRecyclerView = findViewById(R.id.recycler_view);
-        //mRecyclerView.setHasFixedSize(true);
+        mRecyclerView.setHasFixedSize(true);
         mRecyclerView.setLayoutManager(new LinearLayoutManager(this));
 
         //mProgressCircle = findViewById(R.id.progress_circle);
@@ -56,6 +52,9 @@ public class RecyclerActivity extends AppCompatActivity {
             @Override
             public void onDataChange(DataSnapshot dataSnapshot) {
                 for (DataSnapshot postSnapshot : dataSnapshot.getChildren()) {
+//                    if(postSnapshot.getValue() instanceof String){
+//                        continue;
+//                    }
                     Upload upload = postSnapshot.getValue(Upload.class);
                     mUploads.add(upload);
                 }
