+++ app/src/main/java/gcsenxmk/q/SignupActivity.java
@@ -6,7 +6,6 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
-import java.sql.SQLException;
 
 public class SignupActivity  extends AppCompatActivity {
