--- app/src/main/res/layout/activity_createqueue.xml
@@ -1,240 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="fill_parent"
-    android:layout_height="fill_parent"
-    android:orientation="vertical"
-
-    >
-
-
-
-    <LinearLayout
-        android:layout_width="fill_parent"
-        android:layout_height="wrap_content"
-        android:layout_weight="0"
-        android:orientation="horizontal"
-
-        >
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:text="Text"
-            android:id="@+id/textviewemailmerchant"
-            android:layout_gravity="center_horizontal"/>
-
-
-
-
-    </LinearLayout>
-
-    <LinearLayout
-        android:layout_width="fill_parent"
-        android:layout_height="wrap_content"
-        android:layout_weight="0"
-        android:orientation="horizontal"
-
-        >
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:text="Queue"/>
-
-        <EditText
-            android:id="@+id/queuenamemerchant"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-
-            android:layout_weight="3"
-            android:hint="Enter queue name"
-
-
-            />
-
-
-    </LinearLayout>
-
-    <LinearLayout
-        android:layout_width="fill_parent"
-        android:layout_height="wrap_content"
-        android:layout_weight="0"
-        android:orientation="horizontal"
-
-        >
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:text="Description" />
-
-        <EditText
-            android:id="@+id/queuedescription"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-
-            android:layout_weight="3"
-            android:hint="Description of queue"
-
-
-            />
-
-
-    </LinearLayout>
-
-    <LinearLayout
-        android:layout_width="fill_parent"
-        android:layout_height="wrap_content"
-        android:layout_weight="0"
-        android:orientation="horizontal"
-
-        >
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:text="ImageName" />
-
-        <EditText
-            android:id="@+id/edit_text_file_name"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-
-
-            android:layout_toEndOf="@+id/button_choose_image"
-            android:hint="Enter file name" />
-
-
-
-
-    </LinearLayout>
-
-    <LinearLayout
-        android:layout_width="fill_parent"
-        android:layout_height="wrap_content"
-        android:layout_weight="0"
-        android:orientation="horizontal"
-
-        >
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:text="Average waiting time" />
-
-        <EditText
-            android:id="@+id/average_time"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-
-            android:hint="Enter average waiting time" />
-
-
-
-
-    </LinearLayout>
-
-
-
-
-    <LinearLayout
-        android:layout_width="fill_parent"
-        android:layout_height="wrap_content"
-        android:gravity="center"
-        android:orientation="horizontal">
-
-        <Button
-            android:id="@+id/button_choose_image"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Choose file" />
-
-
-
-        <Button
-            android:id="@+id/btnaddqueue"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Create queue"
-            android:layout_gravity="center_horizontal"/>
-
-
-        <Button
-            android:id="@+id/btnLogoutMerchant"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Logout"
-            android:layout_gravity="center_horizontal"/>
-
-
-
-
-
-
-
-
-    </LinearLayout>
-
-
-
-    <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        xmlns:tools="http://schemas.android.com/tools"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:padding="16dp"
-        tools:context="com.codinginflow.firebaseuploadexample.FirebaseUploadActivity">
-
-
-
-
-
-        <ImageView
-            android:id="@+id/image_view"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_above="@id/progress_bar"
-            android:layout_below="@id/edit_text_file_name"
-            android:layout_marginTop="16dp"
-            tools:ignore="NotSibling" />
-
-        <ProgressBar
-            android:id="@+id/progress_bar"
-            style="@style/Widget.AppCompat.ProgressBar.Horizontal"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_above="@+id/button_upload"
-            android:layout_alignParentStart="true"
-            android:layout_marginBottom="16dp" />
-
-        <Button
-            android:id="@+id/button_upload"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_alignParentBottom="true"
-            android:layout_centerHorizontal="true"
-            android:text="Upload" />
-
-        <TextView
-            android:id="@+id/text_view_show_uploads"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_alignBottom="@+id/button_upload"
-            android:layout_alignTop="@+id/button_upload"
-            android:layout_marginStart="25dp"
-            android:layout_toEndOf="@+id/button_upload"
-            android:gravity="center"
-            android:text="Show Uploads"
-            android:textSize="16sp" />
-
-    </RelativeLayout>
-</LinearLayout>
-
